requirements.txt
webvtt-py
ffmpeg